storage, cdi-import: move single-consumer fixtures out of shared tests/conftest.py
vm_instance_from_template_multi_storage_scope_function and
started_windows_vm are only used by
test_successful_vm_from_imported_dv_windows in
tests/storage/cdi_import/test_import_http.py, so they belong in that
feature's own conftest.py rather than the shared tests/conftest.py.
