+++ tests/storage/cdi_import/conftest.py
@@ -23,11 +23,11 @@ from utilities.constants import Images
 from utilities.constants.images import OS_FLAVOR_FEDORA
 from utilities.constants.networking import LINUX_BRIDGE
 from utilities.constants.storage import REGISTRY_STR
-from utilities.constants.timeouts import TIMEOUT_1MIN
+from utilities.constants.timeouts import TIMEOUT_1MIN, TIMEOUT_50MIN
 from utilities.infra import NON_EXIST_URL
 from utilities.network import network_device, network_nad
 from utilities.storage import construct_datavolume_source_dict, create_dv, sc_volume_binding_mode_is_wffc
-from utilities.virt import VirtualMachineForTests
+from utilities.virt import VirtualMachineForTests, running_vm, vm_instance_from_template, wait_for_windows_vm
 
 LOGGER = logging.getLogger(__name__)
 BRIDGE_NAME = "br1-dv"
@@ -224,3 +224,38 @@ def importer_pod_annotations(admin_client, namespace, linux_nad):
         importer_pod = wait_dv_and_get_importer(dv=dv, admin_client=admin_client)
         wait_for_multus_network_status(importer_pod=importer_pod)
         yield importer_pod.instance.metadata.annotations
+
+
+@pytest.fixture()
+def vm_instance_from_template_multi_storage_scope_function(
+    request,
+    unprivileged_client,
+    namespace,
+    data_volume_multi_storage_scope_function,
+    cpu_for_migration,
+):
+    """Calls vm_instance_from_template contextmanager
+
+    Creates a VM from template and starts it (if requested).
+    """
+
+    with vm_instance_from_template(
+        request=request,
+        unprivileged_client=unprivileged_client,
+        namespace=namespace,
+        existing_data_volume=data_volume_multi_storage_scope_function,
+        vm_cpu_model=(cpu_for_migration if request.param.get("set_vm_common_cpu") else None),
+    ) as vm:
+        yield vm
+
+
+@pytest.fixture()
+def started_windows_vm(
+    request,
+    vm_instance_from_template_multi_storage_scope_function,
+):
+    running_vm(vm=vm_instance_from_template_multi_storage_scope_function, dv_wait_timeout=TIMEOUT_50MIN)
+    wait_for_windows_vm(
+        vm=vm_instance_from_template_multi_storage_scope_function,
+        version=request.param["os_version"],
+    )
